Single threaded controller done, gotta fix some crashing issue

+++ driverless_ws/src/controls/src/constants.hpp
@@ -86,7 +86,7 @@ namespace controls {
     constexpr float lookahead_behind_squared = 25.0f;
 
     // AIM communication stuff
-    constexpr float aim_signal_period_ms = 10.0f;
+    constexpr int aim_signal_period_ms = 10.0f;
 
 
     // State Estimation
